Add files via upload
In this project we made use of OOPS and Data Structures. We used concepts like  pointers, array and Linked list, time lapse. We take the details of students one by one and then add them in an Excel file for record keeping. We also made editing functions in case of wrong entry or any correction is done afterwards. Detail of students were printed by entering their ID's. SPI of the students were also shown. We also made a function showing who is at the top of the class and which students get back-log.